Use self-pipe instead of signal-blocking for SIGCHLD/SIGWINCH
pselect() is broken on macOS. This isn't officially documented
anywhere but there are hints of it in this article:

https://daniel.haxx.se/blog/2016/10/11/poll-on-mac-10-12-is-broken/

To safely handle SIGCHLD/SIGWINCH without blocking them,
use the "self-pipe" trick, i.e. write to a pipe when those signals
occur and select() on the read-end of that pipe.

Fixes #19

(cherry picked from commit 1f1ed664d64603f3f1ce1388571227dc723901b2)